refactor(config): rename OpenAPI env vars for billing/company clarity
Rename OPENAPI_* to OPENAPI_BILLING_* for SDI billing vars and
COMPANY_API_* to OPENAPI_COMPANY_* for company lookup vars.
OPENAPI_SANDBOX_MODE stays shared across both modules.

This disambiguates billing and company API configs which both
use the same OpenAPI.com platform but serve different purposes.

Also includes remaining company lookup module frontend/backend
wiring (routes, navigation, API client improvements).

+++ backend/internal/company/models/dto.go
@@ -6,7 +6,7 @@ type OpenAPIBaseResponse struct {
 	Success bool                `json:"success"`
 	Data    *OpenAPICompanyData `json:"data"`
 	Message string              `json:"message,omitempty"`
-	Error   int                 `json:"error,omitempty"`
+	Error   float64             `json:"error,omitempty"`
 }
 
 // OpenAPICompanyData represents the company data from the API response
